Restructure auth feature with clean architecture
