+++ app/(protected)/_layout.tsx
@@ -6,8 +6,8 @@ import { Stack } from "expo-router";
 import { BottomMenu } from "@/components/navigation/bottom-menu/BottomMenu";
 import { Header } from "@/components/navigation/header/Header";
 import { SideMenu } from "@/components/navigation/side-menu/SideMenu";
-import { useSessionContext } from "@/feature/auth/contexts/SessionContext";
-import { useAuthRedirect } from "@/feature/auth/hooks/useAuthRedirect";
+import { useSessionContext } from "@/feature/auth/application/contexts/SessionContext";
+import { useAuthRedirect } from "@/feature/auth/application/hooks/useAuthRedirect";
 import { LinkInputModalProvider } from "@/feature/links/application/context/LinkInputModalContext";
 import { LinkInputView } from "@/feature/links/presentation/views/LinkInputView";
 import { ProfileEditModalProvider } from "@/feature/user/contexts/ProfileEditModalContext";
