+++ app/__tests__/routes/public.test.tsx
@@ -2,7 +2,7 @@ import { Text } from "react-native";
 import { useRouter, useSegments } from "expo-router";
 import { act, render } from "@testing-library/react-native";
 
-import { SessionProvider } from "@/feature/auth/contexts/SessionContext";
+import { SessionProvider } from "@/feature/auth/application/contexts/SessionContext";
 
 // モックの設定
 jest.mock("expo-router", () => ({
