+++ feature/auth/presentation/components/guard/AuthRedirectGuard.tsx
@@ -1,7 +1,7 @@
 import { type ReactNode } from "react";
 
-import { useSessionContext } from "../../contexts/SessionContext";
-import { useAuthRedirect } from "../../hooks/useAuthRedirect";
+import { useSessionContext } from "@/feature/auth/application/contexts/SessionContext";
+import { useAuthRedirect } from "@/feature/auth/application/hooks/useAuthRedirect";
 
 type Props = {
   children: ReactNode;
